requirements.txt
pycups>=2.0.1
Pygments>=2.18.0
WeasyPrint>=62.3
PyPDF2>=3.0.1
reportlab>=4.2.2
redis
python-dotenv
requests
